Fix Token methods & Add null skip token in parser.
Error recovery for parsing is important, Helps to cut the chain of error
you might get. Example: `func void foo() { .... ;   var x: int = 10;`
here after seeing semicolon we expect } r_brace, However , There's no
`}` and currentToken will still be `;` then the parser will not
terminate as no runtime exception used. Move on to var but currentToken
is still `;` as a result we will have chain of error as parseEat will
output so many error til we break and terminate.
SkipTill will give us good start to recover from parsing error and
determine when to terminate as appropriate, However, To implement so
there should be change in almost all parsing implementaion methods and
technieques , And that's a high risk.
This acts like Doc.